FCT accès réseau progress + logout
configuration de l'update progress lorsqu'on veut changer une tache et ajout du logout pour le user dans le navigation bar

+++ tp1/lib/transfer.dart
@@ -45,7 +45,7 @@ class TaskDetailResponse {
 
   int id = 0;
   String name = '';
-  int percentageDone = 0;
+  double percentageDone = 0;
   double percentageTimeSpent = 0.0;
   DateTime deadline = DateTime.now();
   // List<ProgressEvent> events = [];
